backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
psycopg2-binary==2.9.9

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
pylatexenc==2.10
PyMuPDF==1.23.7

# AI and LLM
google-generativeai==0.3.2
openai==1.3.7

# TTS
gTTS==2.5.0
edge-tts==6.1.9

# Image generation
Pillow==10.1.0
reportlab==4.0.7

# Video processing
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
pydantic==2.5.2
pydantic-settings==2.1.0
uuid==1.30
